Helpers for data management

These functions help project to prepare data for long-term storage with descriptive filenames.

TwPrototypes.get_backup_path — Function

get_backup_path(;
    backup_dir="output", date = Dates.Date(Dates.now()), version = "1", extension="csv", 
    project, data_category, experiment)

Create a expressive filepath to store output of experiments.

Arguments

project: project identifier

data_category: kind of data in the project

version: version of the dataset

experiment: further informatin such as scenario, treatment, parameters, etc

extension: file extension

project, datacategory, and version must not contain an underscore, '' to be parsed again by parse_backup_path.

TwPrototypes.parse_backup_path — Function

parse_backup_path(path)

Extract the components from path created by get_backup_path as a NamedTuple.