Makie Helpers

Unit conversions

Many graphical properties are given in pixel or inches. Working with figures for international journals, however, usually requires thinking and specification in metric units. The following function help to convert.

TwPrototypes.cm2inch — Function

cm2inch(x)
cm2inch(x1, x2, x...)

Convert length in cm to inch units: 1 inch = 2.54 The single argument returns a value, the multiple argument version a Tuple. Similar to more verbose syntax using Unitful.jl: uconvert(u"inch",x * u"cm").val.

Figures often look well, if the ratio between length and hight corresponds to the Golden ratio. So we provide it here.

TwPrototypes.golden_ratio — Constant

golden_ratio ≈ 1.618

Two quantities are in the golden ratio if their ratio is the same as the ratio of their sum to the larger of the two quantities.

Makie Layout switching

TwPrototypes.MakieConfig — Type

MakieConfig

A collection of figure properties to tailor the same figure either for presentation in a paper or a presentation.

ppt_MakieConfig(;
    target = :presentation, 
    filetype = "png", 
    fontsize=18, 
    size_inches = (5.0,5.0/golden_ratio), 
    kwargs...) 

paper_MakieConfig(;
    size_inches = cm2inch.((8.3,8.3/golden_ratio)), 
    kwargs...)

Properties and defaults

pt_per_unit = 0.75

filetype = "png"

fontsize = 9 (point units)

size_inches = cm2inch.((17.5,17.5/golden_ratio)) (witdth x height)

TwPrototypes.save_with_config — Function

save_with_config(filename, fig; makie_config = MakieConfig(), args...)

Save figure with file updated extension cfg.filetype to subdirectory cfg.filetype of given path of filename. Set pt_per_unit to makie_config.pt_per_unit.

Shortcuts

TwPrototypes.hidexdecoration! — Function

hidexdecoration!(ax;, label, ticklabels, ticks, grid, minorgrid, minorticks; kwargs...)
hideydecoration!(ax;, label, ticklabels, ticks, grid, minorgrid, minorticks; kwargs...)

Versions of hidexdecorations! and hideydecorations! with defaults reversed. This allows to selectively hide single decorations, e.g. only the label.

TwPrototypes.axis_contents — Function

Extract axis object from given figure position. Works recursively until an Axis-object is returned

Tailored plots

TwPrototypes.density_params — Function

density_params(chns, pars=names(chns, :parameters); 
    makie_config::MakieConfig=MakieConfig(), 
    fig = pdf_figure(cm2inch.((8.3,8.3/1.618)); makie_config), 
    column = 1, xlims=nothing, 
    labels=nothing, colors = nothing, ylabels = nothing, normalize = false, 
    kwargs_axis = repeat([()],length(pars)), 
    prange = (0.025, 0.975), # do not extend x-scale to outliers
    kwargs...
)

Plot density from MCMCChains.value with facets for each parameter.